Work on d_resorce, d_vibration, d_tresure and others (#347)
* Work on d_resorce

* Work on d_a_tag_TWgate

* Work on d_meter_map

* d_tresure OK

* Work on d_vibration

* Work on d_a_npc

* Work on d_a_npc

* Work on d_a_npc

+++ src/d/s/d_s_name.cpp
@@ -109,7 +109,7 @@ dSn_HIO_c::dSn_HIO_c() {
 /* 802587A4-80258820 2530E4 007C+00 1/0 0/0 0/0 .text            phase_1__FPc */
 static s32 phase_1(char* resName) {
     mDoAud_bgmStart(-1);
-    if (dComIfG_setObjectRes(resName, 0, NULL) == 0) {
+    if (dComIfG_setObjectRes(resName, (u8)0, NULL) == 0) {
         return 5;
     }
